% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_editor_fonts.R
\name{add_editor_fonts}
\alias{add_editor_fonts}
\title{add my custom fonts to the editor selection}
\usage{
add_editor_fonts(dash = TRUE)
}
\arguments{
\item{dash}{logical, default TRUE. If TRUE uses .config file on DASH RStudio
server, else uses windows.}
}
\value{
fonts added to UI
}
\description{
Adds my custom fonts (all open source) to the editor UI
}
\author{
Josh Moatt
}
