% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_proj_files.R
\name{add_proj_files}
\alias{add_proj_files}
\title{Add default project files to an existing project/repo}
\usage{
add_proj_files(
  path = here::here(),
  title = "Project title",
  readme = "github",
  gitignore = TRUE
)
}
\arguments{
\item{path}{string containing file path to add files. Default is current
directory (set using \code{here()})}

\item{title}{string with project title. If no title provided, defaults to
"Project title"}

\item{readme}{string specifying readme format. One of "github", "markdown" or
"html". Default is "github"}

\item{github}{logical controlling whether a gitignore is added. Default is
TRUE}
}
\value{
project files added to specified directory
}
\description{
Simple function to add the standard project files to an existing
R Project or local GitHub Repo. This should be used when setting up a new
project after initial set up - particularly useful for if you have just
cloned a new GitHub repo and want to add the template project files.
}
\details{
The function is designed to be used after manually creating a local
version of a GitHub repo.

The function will add the following:

\itemize{
\item data, src and outputs folders
\item README using \code{create_fs_readme}
\item pipeline script using \code{create_fs_script}
\item optionally creates a gitignore using \code{create_fs_gitignore}
}

The function allows users to include the author and project title when
calling the function, which will be prefilled
}
\author{
Josh Moatt
}
