% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_themes.R
\name{add_themes}
\alias{add_themes}
\title{Add my rstudio themes}
\usage{
add_themes()
}
\value{
themes added to the rstudio api
}
\description{
Function to add my rstudio themes from within this package.
}
\author{
Josh Moatt
}
