% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_github_ssh.R
\name{connect_github_ssh}
\alias{connect_github_ssh}
\title{Connect RStudio to your GitHub account using a SSH}
\usage{
connect_github_ssh()
}
\description{
Connect RStudio to your GitHub account using SSH - this is the
preferred method of connecting RStudio and GitHub on the DASH platform. The
function is set up to use my username and email address,using these to set
my credentials and generate an SSH key. Once the SSH key has been added to
GitHub, the function does the final bits of set-up. All steps and outputs
are printed in the RStudio console rather than the terminal, which offers a
more user friendly experience.
}
\details{
The suggested method for connecting RStudio and GitHub on the DASH
platform is via SSH. The alternative is via Personal Access Token, but this
method can be a bit clunky and frustrating on the DASH platform. With SSH
providing a much smoother user experience.

This function is designed to simplify the connection process and avoid the
me having to interact with the terminal. Instead, all prompts and outputs
are returned in the RStudio console.

The function will set my username and email then print an output to confirm
they have been set. The function will then create a hidden SSH folder and
sub-folder for "id_ed25519" (~/.ssh/id_ed25519). Once done, it will then
generate an SSH key and save it to this folder. The SSH key will then be
printed in the console. The function then adds GitHub as a known host.

At this point, the function pauses and asks the for confirmation that the
SSH key has been added to GitHub. Now I can then copy the SSH key from the
console, go to GitHub -> settings -> SSH and GPG keys and add the SSH key.

Once added, the I can respond to the prompt and the function will finish
establishing the connection. If all works as expected, the message "Hi
jpmoatt! You've successfully authenticated, but GitHub does not provide
shell access" will be printed in the console.
}
\author{
Josh Moatt
}
