% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_gitignore.R
\name{create_gitignore}
\alias{create_gitignore}
\title{Create a gitignore file based on the my template}
\usage{
create_gitignore(type = "default", file_path = NULL, custom_txt = NULL)
}
\arguments{
\item{type}{description controlling which gitignore is added. "default" will
add the standard template. "custom" will enable the user to provide a
custom template via \code{custom_txt}.}

\item{file_path}{optional argument allowing users to specify file path where
gitignore should be created. If not entered, will default to current
project/working directory.}

\item{custom_txt}{optional argument allowing users to provide their own
gitignore template. Must be provided as a string. Only used if type set to
"custom".}
}
\value{
A gitignore file is added to the project.
}
\description{
Use this function to create a gitignore for a  project.
}
\details{
This function will create a gitignore for a project using a
pre-defined template.

Used as default, it will automatically add the data and output folders (as
created in the project template). This is to ensure nor restricted data or
unpublished results are accidentally pushed to GitHub.

Alternatively, a custom gitignore can be provided by setting \code{type} to
"custom" and providing a custom template to \code{custom_txt} as a string.

Note it will replace any existing .gitignore files present in the project
already.
}
\author{
Josh Moatt
}
