% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_script.R
\name{create_script}
\alias{create_script}
\title{Create a new script with my header.}
\usage{
create_script(
  file_name = NULL,
  file_path = NULL,
  project = NULL,
  date = format(Sys.Date(), "\%d/\%m/\%Y")
)
}
\arguments{
\item{file_name}{string containing desired name for script.}

\item{file_path}{string containing folder name to save script. This is built
on the here function in R, so will follow your root directory. If you want
to save in a sub-folder, enter the full folder sequence, e.g.
"folder/sub-folder".}

\item{date}{string containing a date. By default, this will be set as today's
date.}
}
\value{
An R script will be saved in the root directory or in the specified
folder.
}
\description{
This function will create a new script with my header added.
}
\author{
Josh Moatt
}
