% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_addin.R
\name{script_addin}
\alias{script_addin}
\title{Addin to open new script with my template}
\usage{
script_addin()
}
\value{
A script will open in RStudio.
}
\description{
This addin will open a new blank R script which is populated
with my header.
}
\details{
This addin will create a new blank script that is populated with my
template header.

Addins must use the rstudioapi package. This addin uses the \code{\link[=documentNew]{documentNew()}}
function from the \code{rstudioapi} package to open the new R script.

Note: this opens an normal untitled script. It must be manually saved in
the correct directory.
}
\author{
Josh Moatt
}
