% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_addins.R
\name{sqllight_addin}
\alias{sqllight_addin}
\alias{kisstheme_addin}
\alias{draculatheme_addin}
\alias{cobalttheme_addin}
\title{Apply my rstudio themes via addins}
\usage{
sqllight_addin()

kisstheme_addin()

draculatheme_addin()

cobalttheme_addin()
}
\value{
themes added to the rstudio api
}
\description{
Addins to apply my rstudio themes
}
\author{
Josh Moatt
}
