% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_github.R
\name{use_github}
\alias{use_github}
\title{Create a GitHub repository from local project.}
\usage{
use_github(message = "Initial commit", private = TRUE)
}
\arguments{
\item{message}{initial commit message. Default is "Initial commit".}

\item{private}{if TRUE creates private repo.}
}
\value{
R project is truned into a git repo and an associated github repo si created
in the users github account.
}
\description{
This simple function will turn your local R project into a git repo, then
create a repo on GitHub and perform the initial set up and commit.

Note: by default creates a private repo to ensure security. Can be change if
the repo needs to be public. Can also be done in GitHub at a later date.
}
\details{
This function will take an existing R project on you local machine, turn it
into a git repo and create a GitHub repository. It will do the following:

\itemize{
\item add a gitignore
\item initalise the git repo
\item stage any uncommitted files
\item commit the files
\item create a GitHub repo
\item restart RStudio to activate the git pane in R
}

The gitignore is added using the \code{\link[=useful_gitignore]{useful_gitignore()}} function within this
package.

The repo initialisation, staging and committing is all done using the \code{gert}
package.

Creating the GitHub repo uses the \code{\link[=use_github]{use_github()}} function from the \code{usethis}
package.

For this function to work you must:

\itemize{
\item have git installed on your machine
\item have a GitHub account
\item have your GitHub credentials entered into RStudio
\item have your Personal Access Token (PAT) entered in RStudio
}

Note: occasionally the function seems to fails to set the master branch and
users are unable to push changes. If this happens try running \verb{git push -u origin master} in the terminal, this should set you current branch as the
master. We're not sure why this happens, but it is advisable to use
\code{\link[=connect_github]{connect_github()}} to set your credentials properly before trying this
function.
}
