% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_add_themes.R
\name{useful_add_themes}
\alias{useful_add_themes}
\title{Add my rstudio themes}
\usage{
useful_add_themes()
}
\value{
themes added to the rstudio api
}
\description{
Function to add my rstudio themes from within this package.
}
\author{
Josh Moatt
}
