% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_connect_github.R
\name{useful_connect_github}
\alias{useful_connect_github}
\title{Connect RStudio to your GitHub account}
\usage{
useful_connect_github(defra = TRUE)
}
\arguments{
\item{defra}{default TRUE. If TRUE will set the defra proxy in the terminal
(needed for linking to github)}
}
\value{
GitHub credentials and PAT set
}
\description{
This is a simple function which connect RStudio to GitHub, allowing you to
work on GitHub repositories. This is an essential part of Reproducible
Analytical Pipelines and best practice for coding.
}
\details{
This is a simple function which will set you GitHub credentials and Personal
Access Token (PAT) and connect RStudio to GitHub. This is essential if you
want to work in RStudio in projects/repos stored on GitHub.

It uses \code{\link[=system]{system()}} to run the necessary code in the terminal to set your
credentials, and \code{\link[=gitcreds_set]{gitcreds_set()}} and \code{\link[=set_github_pat]{set_github_pat()}} from the \code{gitcreds}
and \code{credentials} packages to connect your RStudio to GitHub.

\code{\link[=gitcreds_set]{gitcreds_set()}} is an interactive function and will prompt users for input.
To replace existing credentials/PAT choose option 2. You will then be
prompted for you PAT. PAT should be changed every 30 days to
ensure security.

Note: For this function to work you must:

\itemize{
\item have git installed on your local machine
\item have a GitHub account
\item have created a Personal Access Token (PAT) on GitHub.
}

Guidance on how to create a PAT can be found here: ADD LINK.
}
