% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_proj_pat.R
\name{useful_proj_pat}
\alias{useful_proj_pat}
\title{Add \code{\link[=set_github_pat]{set_github_pat()}} to project specific .Rprofiles}
\usage{
useful_proj_pat()
}
\value{
updated .Rprofile
}
\description{
A simple function to add the \code{\link[=set_github_pat]{set_github_pat()}} function call to your project
specific .Rprofile. This is needed for projects where \code{renv} is activated, so
you can avoid hardcoding your PAT anywhere. See \link{useful_connect_github} for
full details and reasoning behind this function.
}
\author{
Josh Moatt
}
