% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_proj_template.R
\name{useful_proj_template}
\alias{useful_proj_template}
\title{Function for RStudio project template}
\usage{
useful_proj_template(path, ...)
}
\description{
This is a function that is called in the "New project" viewer pane when the
user chooses the useful project template. It should not be used away from the
RStudio "New project" viewer.

I have not included additional information on how to use this function, as it
is not intended to be used outside the template call.

To subsequently link this to a github repo, the best plan is to use
\code{\link[=fs_use_github]{fs_use_github()}}.
}
\author{
Josh Moatt
}
