% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_readme.R
\name{useful_readme}
\alias{useful_readme}
\title{Create a README using my template}
\usage{
useful_readme(
  format = c("markdown", "github", "html"),
  file_path = NULL,
  readme_title = NULL
)
}
\arguments{
\item{format}{controls the output format of the README. Default is "markdown",
but can be "html" or "github".}

\item{file_path}{sting containing file path where README will be saved.}

\item{readme_title}{string containing README title. If no string provided will
be set to "README (edit title)".}
}
\value{
Output is a .qmd file containing the desired README template and an
initial render of the README in the desired output.
}
\description{
Use this function to create a README for a project.
}
\details{
This function will create a README for a project using a pre-defined
template. The function will create a Quarto (.qmd) file and do a first render
producing a desired output.

The output type can be controlled using the format option (default is
markdown). Can also be html or github (gfm).
}
\author{
Josh Moatt
}
