% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_roxygen_addin.R
\name{useful_roxygen_addin}
\alias{useful_roxygen_addin}
\title{Addin to open new script with roxygen template}
\usage{
useful_roxygen_addin()
}
\value{
A script will open in RStudio.
}
\description{
This addin will open a new blank R script which is populated with a standard
roxygen header.
}
\details{
This addin will create a new blank script that is populated with standard
roxygen header.

Addins must use the rstudioapi package. This addin uses the \code{\link[=documentNew]{documentNew()}}
function from the \code{rstudioapi} package to open the new R script.

Note: this opens an normal untitled script. It must be manually saved in the
correct directory.
}
\author{
Josh Moatt
}
