% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_rstudio_setup.R
\name{useful_rstudio_setup}
\alias{useful_rstudio_setup}
\title{Apply my preffered RStudio settings and layout.}
\usage{
useful_rstudio_setup()
}
\value{
RStudio configured to my settings
}
\description{
This function will change the settings and layout of RStudio to my usual
preferences. This has the source pane on the left, the console and terminal
on the top right and all others on the bottom right. It also applies my usual
global options (E.g. rainbow brackedts, hexcode preview on etc).
}
\author{
Josh Moatt
}
