% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_script_template.R
\name{useful_script_template}
\alias{useful_script_template}
\title{Create an default R script template.}
\usage{
useful_script_template(
  format = c("mine", "custom", "manual_edit", "blank"),
  template = NULL,
  dash = TRUE
)
}
\arguments{
\item{format}{what format the template will take. There are four options:
"mine" (my default template), "custom" will apply a custom template
(provided as a string), "manual_edit" will open the template so you can
manually edit the template, and "blank" can be used to remove all
pre-existing templates. Note: manual edit can also be used to edit existing
templates.}

\item{template}{default is NULL, only used if format = "custom". Used to
provide custom template design. Must be provided as a string. Must be
provided if using format = "custom" or the function will return an error.}

\item{dash}{default is FALSE. If TRUE changes the file path to the one needed
for the RStudio server on DASH}
}
\value{
New .R file created containing the script template
}
\description{
This function can be used to create a default R script template. All scripts
opened from this point will have this template applied as a header.
}
\details{
This function is used to create a new template for R scripts. This will be
the default that all subsequent R scripts opened will contain. By doing this
is it should be easier to follow best practice and properly comment all
scripts you create.

The default is stored in the appdata folder on your c drive:
"~/AppData/Roaming/RStudio". It will create a "templates" folder where the
default will be stored.

The function has various ways it can work which will give you the ability to
create whatever header template you wish. It has four ways formatting
options:

\itemize{
\item mine
\item custom
\item manual_edit
\item blank
}

"mine" will pre-load the default script with my template.

"custom" will allow you to provide your own custom template as a string,
which will then be added to the default.

"manual" will allow you to manually edit the default template. It will open
it in your R studio window and manual edits can be saved. Note, this can also
be used to tweak a pre-existing template (e.g. to add your name and email to
all scripts).

"blank" will delete the default R script and template. This returns R back to
normal, and any script opened subsequently will be blank.
}
\author{
Josh Moatt
}
