% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_addins.R
\name{lighttheme_addin}
\alias{lighttheme_addin}
\alias{kisstheme_addin}
\alias{chaostheme_addin}
\alias{grubbertheme_addin}
\alias{randomtheme_addin}
\title{Apply my rstudio themes via addins}
\usage{
lighttheme_addin()

kisstheme_addin()

chaostheme_addin()

grubbertheme_addin()

randomtheme_addin()
}
\value{
themes added to the rstudio api
}
\description{
Addins to apply my rstudio themes
}
\author{
Josh Moatt
}
