% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_apply_theme_addins.R
\name{useful_lighttheme_addin}
\alias{useful_lighttheme_addin}
\alias{useful_darktheme_addin}
\title{Apply my rstudio themes via addins}
\usage{
useful_lighttheme_addin()

useful_darktheme_addin()
}
\value{
themes added to the rstudio api
}
\description{
Addins to apply my rstudio themes
}
\author{
Josh Moatt
}
